create display file
